+++ examples/contacts/deleteContact.php
@@ -1,0 +1,9 @@
+<?php
+
+use jamesiarmes\PEWS\Contacts\ContactsAPI as API;
+
+$api = API::withUsernameAndPassword('server', 'username', 'password');
+
+$contact = $api->getContacts();
+
+$api->deleteItems($contact[0]->getItemId());